@@ -14,12 +14,20 @@
     "core-js": "^2.4.1",
     "express": "^4.14.0",
     "flux": "^2.1.1",
-    "react": "^0.14.6",
-    "react-autosuggest": "^7.0.1",
-    "react-dom": "^0.14.6",
+    "react": "15.5.4",
+    "react-addons-perf": "15.0.1",
+    "react-addons-test-utils": "15.0.1",
+    "react-addons-update": "15.0.1",
+    "react-dom": "^15.0.1",
+    "react-mixin": "^1.7.0",
+    "react-router": "4.1.1",
+    "react-addons-linked-state-mixin": "15.4.2",
     "webpack": "^1.13.2",
     "webpack-dev-server": "^1.14.1",
-    "history": "^1.17.0"
+    "history": "^1.17.0",
+    "reqwest": "^1.1.5",
+    "when": "^3.7.2",
+    "jwt-decode": "^1.1.0"
     },
   "devDependencies": {
     "classnames": "^2.1.3",
