@@ -1,45 +1,51 @@
 import React, { Component } from 'react';
-import StartPage from './StartPage';
-import EmployeeSessionStore from '../stores/EmployeeSessionStore';
-import * as EmployeeSignInActions from '../actions/EmployeeSignInActions';
+import ReactMixin from 'react-mixin';
+import LinkedStateMixin from 'react-addons-linked-state-mixin';
+import Auth from '../services/AuthService';
 
-var SignIn = React.createClass({
-  //class SignIn extends Component {
-  propTypes: {
-    userId: React.PropTypes.string,
-    showComponent: React.PropTypes.bool,
-  },
+export default class SignIn extends React.Component {
+  constructor() {
+    super()
+    this.state = {
+      userId: ''
+    };
+  }
   getDefaultProps() {
     return {
       userId: '',
       showComponent: false,
     };
-  },
+  }
   getInitialState() {
     return {
       showComponent: false,
       userId: '',
     };
-  },
-
+  }
   OnClickSignIn() {
     this.setState({ showComponent: true, userId: document.getElementById("loginUserId").value });
     EmployeeSessionStore.userLogin(userId);
-  },
-
+  }
   OnClickDisplayNumber(item) {
     var userInput = this.state.userId += item;
     console.log('User input value : ' + item + ' \n Full User ID : ' + this.state.userId);
     this.setState({ userId: userInput });
-  },
-
-  render() {
+  }
+ login(e) {
+    e.preventDefault();
+    Auth.login(this.state.user, this.state.password)
+      .catch(function(err) {
+        alert("There's an error logging in");
+        console.log("Error logging in", err);
+      });
+  }
+/**   render() {
     return (
       <div id="SignInPage" >
         <div id="NumPadDiv" class='SignInNumPad'>
           <h4 class="Heading"> OpenPOS </h4>
           <h1 class="Heading"> Sign In </h1>
-          <input type="password" id="loginUserId" class="loginUserId" value={this.state.userId} ></input><br />
+          <input type="password" id="loginUserId" valueLink={this.linkState('userId')} class="loginUserId"></input><br />
           <button class="NumPad-button" data-char='1' onClick={this.OnClickDisplayNumber.bind(this, '1')}>1</button><span class="NumPad-button" />
           <button class="NumPad-button" data-char='2' onClick={this.OnClickDisplayNumber.bind(this, '2')}>2</button><span class="NumPad-button" />
           <button class="NumPad-button" data-char='3' onClick={this.OnClickDisplayNumber.bind(this, '3')}>3</button><br />
@@ -51,7 +57,7 @@ var SignIn = React.createClass({
           <button class="NumPad-button" data-char='9' onClick={this.OnClickDisplayNumber.bind(this, '9')}>9</button><br />
           <button class="NumPad-button" data-char='0' onClick={this.OnClickDisplayNumber.bind(this, '0')}>0</button>
           <br />
-          <button id='SignInNumPad-Sign' onClick={this.OnClickSignIn.bind(this)}>Sign In</button>
+          <button id='SignInNumPad-Sign' onClick={this.login.bind(this)}>Sign In</button>
           <button id='SignInNumPad-Clear' onClick="">Clear</button>
         </div>
         {
@@ -68,9 +74,25 @@ var SignIn = React.createClass({
 
       </div>
     );
+  }*/
+    render() {
+    return (
+      <div className="login jumbotron center-block">
+        <h1>Login</h1>
+        <form role="form">
+        <div className="form-group">
+          <label htmlFor="username">Username</label>
+          <input type="text" valueLink={this.linkState('user')} className="form-control" id="username" placeholder="Username" />
+        </div>
+        <div className="form-group">
+          <label htmlFor="password">Password</label>
+          <input type="password" valueLink={this.linkState('password')} className="form-control" id="password" ref="password" placeholder="Password" />
+        </div>
+        <button type="submit" className="btn btn-default" onClick={this.login.bind(this)}>Submit</button>
+      </form>
+    </div>
+    );
   }
-});
+}
 
-module.exports = SignIn;
-
-
+ReactMixin(SignIn.prototype, LinkedStateMixin);
