@@ -3,9 +3,9 @@ var webpack = require('webpack');
 var path = require('path');
 
 module.exports = {
-  context: path.join(__dirname, "src"),
+  context: path.join(__dirname, 'src'),
   devtool: debug ? "inline-sourcemap" : null,
-  entry: "./js/client.js",
+  entry: './js/client.js',
   module: {
     loaders: [
       {
